@@ -8,50 +8,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.1",
-    "@angular/common": "~12.0.1",
-    "@angular/compiler": "~12.0.1",
-    "@angular/core": "~12.0.1",
-    "@angular/forms": "~12.0.1",
-    "@angular/platform-browser": "~12.0.1",
-    "@angular/platform-browser-dynamic": "~12.0.1",
-    "@angular/router": "~12.0.1",
-    "@microsoft/signalr": "^5.0.1",
-    "@po-ui/ng-components": "^5.0.0",
-    "@totvs/po-theme": "^5.0.0",
+    "@angular/animations": "~13.2.0",
+    "@angular/common": "~13.2.0",
+    "@angular/compiler": "~13.2.0",
+    "@angular/core": "~13.2.0",
+    "@angular/forms": "~13.2.0",
+    "@angular/platform-browser": "~13.2.0",
+    "@angular/platform-browser-dynamic": "~13.2.0",
+    "@angular/router": "~13.2.0",
+    "@microsoft/signalr": "^6.0.0",
+    "@po-ui/ng-components": "^6.1.0",
+    "@po-ui/style": "^6.1.0",
+    "@totvs/po-theme": "^6.1.0",
     "jwt-decode": "^3.1.2",
-    "ngx-cookie-service": "^3.0.4",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.1.0",
+    "ngx-cookie-service": "^13.1.2",
+    "rxjs": "~7.4.0",
+    "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.1",
-    "@angular-eslint/builder": "12.1.0",
-    "@angular-eslint/eslint-plugin": "12.1.0",
-    "@angular-eslint/eslint-plugin-template": "12.1.0",
-    "@angular-eslint/schematics": "12.1.0",
-    "@angular-eslint/template-parser": "12.1.0",
-    "@angular/cli": "~12.0.1",
-    "@angular/compiler-cli": "~12.0.1",
-    "@angular/language-service": "~11.0.1",
-    "@types/jasmine": "~3.6.0",
+    "@angular-devkit/build-angular": "~13.2.0",
+    "@angular/cli": "~13.2.0",
+    "@angular/compiler-cli": "~13.2.0",
+    "@angular/language-service": "~13.2.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "^4.27.0",
-    "@typescript-eslint/parser": "4.23.0",
-    "eslint": "^7.28.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "jasmine-core": "~3.7.0",
-    "jasmine-spec-reporter": "~5.0.0",
+    "@types/jasmine": "~3.10.0",
+    "jasmine-core": "~3.8.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.2.4"
+    "typescript": "~4.5.5"
   }
 }
