@@ -3,6 +3,7 @@ import { HttpClient } from '@angular/common/http';
 import { environment } from 'src/environments/environment';
 import { SalaParticipante } from '../models/sala-participante.model';
 import { PoNotificationService } from '@po-ui/ng-components';
+import { lastValueFrom } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -17,8 +18,8 @@ export class SalaParticipanteService {
   ) { }
 
   public incluirOuAlterarParticipante(participante: SalaParticipante): Promise<SalaParticipante> {
-    return this.http.post<SalaParticipante>(`${this.url}/Participante/`, participante)
-      .toPromise()
+    const request = this.http.post<SalaParticipante>(`${this.url}/Participante/`, participante)
+    return lastValueFrom(request)
       .catch( erro => {
           this.poNotification.error('Houve um erro ao tentar incluir o participante.');
           console.log(erro);
